Migrate helper function to Utils directory
