+++ app/utils/auth.py
@@ -1,6 +1,8 @@
 from flask import request, jsonify, g
 from functools import wraps
 from firebase_admin import auth as firebase_auth
+from app.models import User
+from app import db, current_app
 
 
 def firebase_auth_required(f):
@@ -32,3 +34,23 @@ def firebase_auth_required(f):
         return f(*args, **kwargs)
 
     return decorated_function
+
+
+def get_or_create_user(firebase_uid: str) -> User:
+    """Retrieve or create a user from User table by Firebase UID.
+
+    Args:
+        firebase_uid: Firebase UID derived from request's header bearer token.
+
+    Returns:
+        user: Retrieved or newly created User instance.
+    """
+    user = User.query.filter_by(firebase_uid=firebase_uid).first()
+    current_app.logger.info(f"Retrieved {user} of Firebase UID {firebase_uid}.")
+
+    if not user:
+        user = User(firebase_uid=firebase_uid)
+        db.session.add(user)
+        db.session.commit()
+        current_app.logger.info(f"Created {user} of Firebase UID {firebase_uid}.")
+    return user
